eina: Set it to NULL before next iteration
it might have been free'd by the user, so set it to NULL before next
iteration. This is an attempt to fix CID 1039913 and 1039914.

We don't use the pointer value, only the pointer, so the error is wrong.
Could flag the error in coverity, but if this fixes it, we wont see the
error in other situations.

Backport from commit 08978ddb560ad10e4427fc7b4aaa18cf6912ba0c
on https://git.enlightenment.org/core/efl.git