Fix very strange behavior when listing long directories.
The removed code triggers a very annoying bug :
If you list a directory whose fullpath is very long, some of its files
might seem to appear twice.
This is dues to this MAX_PATH check, that will replace the name of the
file by the name of the next file (which COULD be also too long).

Here, it happened by declaring us two files with the exact same name,
one being a file, the other being a directory, leading to some big bugs.